branchmap {denpro}	R Documentation

Computes a branching map from a sequence of level set trees

Description

Branching map visualizes the levels of branching of level set trees of estimates belonging to a scale of estimates. It visualizes also the excess masses of the roots of the branches.

Usage

branchmap(estiseq, hseq = NULL, levnum = 80, paletti = NULL, 
rootpaletti = NULL, type = "jump")

Arguments

estiseq
A sequence of estimates and level set trees of the estimates. Output of function lstseq.kern or function lstseq.carthisto.

hseq
The sequence of smoothing parameters of the scale of estimates.

levnum
The number of level sets used to approximate the level set trees.

paletti
A sequence of color names; colors for each branch, other than the root branches.

rootpaletti
A sequence of color names; colors for the root branches.

type
internal

Value

A representation as a list of a 2D function

level
x-coordinate is the level of the level sets

h
y-coordinate is the smoothing parameter

z
z-coordinate is the excess mass

col
colors for the graph of the 2D function

Author(s)

Jussi Klemela

See Also

lstseq.kern, plotbranchmap

Examples

dendat<-sim.data(n=200,type="mulmod")

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

N<-c(16,16)
estiseq<-lstseq.kern(dendat,hseq,N,lstree=TRUE)

bm<-branchmap(estiseq)

plotbranchmap(bm)

[Package denpro version 0.9.2 Index]