excmas {denpro}	R Documentation

Computes the excess masses associated with the nodes of a level set tree

Description

Given a level set tree, calculates the excess mass for every node of the tree. The excess mass is the probability mass under the curve, over a given level, in a certain branch of the level set tree. For example, function "leafsfirst" returns a level set tree.

Usage

excmas(lst)

Arguments

lst
level set tree object; list of vectors which is the output of for example "leafsfirst", "profkern", "profhist", or "profgene"

Value

Vector whose length is equal to the number of nodes of the level set tree.

Author(s)

Jussi Klemela

See Also

leafsfirst, plottree

Examples

set.seed(1)
dendat<-matrix(rnorm(20),10)   #10*2 data-matrix
pcf<-pcf.kern(dendat,h=1,N=c(16,16))
lst<-leafsfirst(pcf)
excmas(lst)

[Package denpro version 0.9.2 Index]