exmap {denpro}	R Documentation

Computes a scale of excess mass profiles

Description

Calculates a scale of excess mass profiles from a scale of estimates given as level set trees

Usage

exmap(estiseq, mt, hind = NULL, hseq = NULL, n = NULL, moteslist = NULL, 
ylist = NULL)

Arguments

estiseq
a scale of estimates, output for example of "lstseq.kern"

mt
mode graph; output of "modegraph"

hind
the indexes of the estimates for which the excess mass profiles will be calculated

hseq
the sequence of smoothing parameters corresponding to estimates

n
sample size

moteslist
internal

ylist
internal, alternative to hind

Value

A list of excess mass profiles

Author(s)

Jussi Klemela

See Also

plotexmap, modegraph

Examples

dendat<-sim.data(n=200,type="mulmod")

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

N<-c(16,16)
estiseq<-lstseq.kern(dendat,hseq,N,lstree=TRUE)

mt<-modegraph(estiseq)

horilines<-c(1,2,3,4)

sp<-exmap(estiseq,mt,horilines)

plotexmap(sp,mt,ylim=c(1,4.5))

[Package denpro version 0.9.2 Index]