hgrid {denpro}	R Documentation

Returns a grid of smoothing parameter values

Description

Returns a grid of smoothing parameter values for kernel estimates, either with a logarithmic spacing, or with equal spacing, in decreasing order.

Usage

hgrid(h1, h2, lkm, base = 10)

Arguments

h1
the lowest smoothing parameter value

h2
the largest smoothing parameter value

lkm
the number of smoothing parameters in the grid

base
the base of the logarithm, used in the logarithmic spacing, if NULL, then the equal spacing is applied

Value

a vector of smoothing parameters

Author(s)

Jussi Klemela

See Also

lstseq.kern, branchmap, modegraph, exmap

Examples

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

hseq<-hgrid(h1,h2,lkm,base=NULL)

[Package denpro version 0.9.2 Index]