locofmax {denpro}	R Documentation

Calculates the location of the maximum of a function

Description

Calculates the location of the maximum of a piecewise constant function.

Usage

locofmax(pcf)

Arguments

pcf
piecewise constant function; output of "pcf.kern" or "pcf.func"

Value

a d-vector; the location of the maximum; this is defined as the barycenter of the set where the function has the maximum

Author(s)

Jussi Klemela

See Also

pcf.kern, pcf.func

Examples

dendat<-sim.data(n=100,type="mulmod")
pcf<-pcf.kern(dendat,h=1,N=c(32,32))

locofmax(pcf)

[Package denpro version 0.9.2 Index]