lstseq.kern {denpro}	R Documentation

Calculates a scale of kernel estimates

Description

Calculates a scale of kernel estimates corresponding to a scale of smoothing parmeters.

Usage

lstseq.kern(dendat, hseq, N, lstree = NULL, level = NULL,
Q = NULL, kernel = "gauss", hw = NULL, algo = "leafsfirst", support = NULL)

Arguments

dendat
n*d matrix of real numbers; the data matrix

hseq
a vector of positive real numbers; the sequence should be monotonic

N
vector of d positive integers; the dimension of the grid where the kernel estimate will be evaluated; we evaluate the estimate on a regular grid which contains the support of the kernel estimate

lstree
if NULL, then level set trees are not calculated

level
NULL or a real number between 0 and 1; if NULL, then shape trees are not calculated; if number, then it is the level in percents of the maximum of the level sets for which the shape trees are calculated

Q
positive integer; needed only in the DynaDecompose algorithm, see parameter "algo"; the number of levels in the level set trees

kernel
"epane" or "gauss"; the kernel is either the Bartlett-Epanechnikov product kernel or the standard Gaussian

hw
positive integer; parameter for time localized kernel estimation; gives the smoothing parameter for the temporal smoothing

algo
"leafsfirst" or "dynadecompose"

support
2*d vector of reals gives the d intervals of a rectangular support; c(low1,upp1,...,lowd,uppd)

Value

A list with components

lstseq
a list of level set trees

pcfseq
a list of piecewise constant functions

stseq
a list of shape trees

hseq
a vector of smoothing parameters corresponding to the members in the sequences

Author(s)

Jussi Klemela

See Also

scaletable

Examples

dendat<-sim.data(n=200,type="mulmod")

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

N<-c(16,16)
estiseq<-lstseq.kern(dendat,hseq,N)

[Package denpro version 0.9.2 Index]