modegraph {denpro}	R Documentation

Calculates a mode graph from a scale of estimates

Description

Calculates a mode graph from a scale of estimates. The estimates in the scale are given as level set trees.

Usage

modegraph(estiseq, hseq = NULL, paletti = NULL)

Arguments

estiseq
a scale of estimates, output for example of "lstseq.kern"

hseq
the corresponding scale of smoothing parameters

paletti
vector of color names

Value

mode graph

Author(s)

Jussi Klemela

See Also

lstseq.kern, plotmodet

Examples

dendat<-sim.data(n=200,type="mulmod")

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

N<-c(16,16)
estiseq<-lstseq.kern(dendat,hseq,N,lstree=TRUE)

mt<-modegraph(estiseq)

plotmodet(mt)

[Package denpro version 0.9.2 Index]