plotexmap {denpro}	R Documentation

Plots a scale of excess mass profiles

Description

Plots a scale of excess mass profiles.

Usage

plotexmap(sp, mt, xaxt = "n", lift = 0.1, leaflift = 0.1, ylim = NULL, 
leafcolors = NULL)

Arguments

sp
scale of excess mass profiles; output of "exmap"

mt
mode graph; output of "modegraph"

xaxt
"n" or "y"; whether x-axis will be plotted

lift
positive real number; space between the lines of the excess mass profiles

leaflift
positive real number; lifting for the bullets associated with the leaf nodes of the excess mass profiles

ylim
2-vector; gives the interval of the y-axis

leafcolors
a paletti of the colors for the labeling of the leaf nodes

Value

Plots the scale of excess mass profiles to the graphic device

Author(s)

Jussi Klemela

See Also

exmap, modegraph

Examples

dendat<-sim.data(n=200,type="mulmod")

h1<-0.9
h2<-2.2
lkm<-5
hseq<-hgrid(h1,h2,lkm)

N<-c(16,16)
estiseq<-lstseq.kern(dendat,hseq,N,lstree=TRUE)

mt<-modegraph(estiseq)

horilines<-c(1,2,3,4)

sp<-exmap(estiseq,mt,horilines)

plotexmap(sp,mt,ylim=c(1,4.5))

[Package denpro version 0.9.2 Index]