stseq {denpro}	R Documentation

Calculates a sequence of radius functions from a sequence of level sets

Description

Calculates a sequence of radius functions from level sets of a function of a given parametric form.

Usage

stseq(N, lnum, refe = NULL, func = NULL, dendat = NULL, h = NULL, Q = NULL, 
kernel = "epane", weights = NULL, sig = rep(1, length(N)), support = NULL, 
theta = NULL, M = NULL, p = NULL, mul = 3, t = rep(1, length(N)), 
marginal = "normal", r = 0, mu = NULL, xi = NULL, Omega = NULL, alpha = NULL, 
df = NULL, g = 1, base = 10)

Arguments

N
d-vector of integers >1; the size of the grid where the function is evaluated

lnum
positive integer; the number of level sets from which the transforms are calculated

refe
d-vector of real numbers; the reference point for the shape trees; if refe=NULL, then the location of the maximum of the function will be used

func
character; the name of the function, see the "pcf.func" for the possibilities

dendat
n*d matrix of data; when the function is a kernel estimate one gives the data as argument

h
postive real number; the smoothing parameter of the kernel estimate

Q
internal

kernel
"gauss" or "epane"; the kernel of the kernel estimate

weights
n vector of nonnegative weights, where n is the sample size; sum of the elements of "weights" should be one; these are the weights of a time localized kernel estimator

sig
see "pcf.func"

support
see "pcf.func"

theta
see "pcf.func"

M
see "pcf.func"

p
see "pcf.func"

mul
see "pcf.func"

t
see "pcf.func"

marginal
see "pcf.func"

r
see "pcf.func"

mu
see "pcf.func"

xi
see "pcf.func"

Omega
see "pcf.func"

alpha
see "pcf.func"

df
see "pcf.func"

g
see "pcf.func"

base
positve integer or NULL; the base of the logarithm, when the logarithmic spacing is used for the level sets, if "base" is NULL, then the level sets are equispaced

Value

A list of radius transforms.

Author(s)

Jussi Klemela

See Also

shape2d, plotvolu2d

Examples

N<-c(32,32)
lnum<-30
func<-"prod"
marginal<-"student"
yla<-5
support<-c(-yla,yla,-yla,yla)
g<-c(1,3)
st<-stseq(N,lnum,func=func,marginal=marginal,support=support,g=g)

gnum<-70  
vd<-shape2d(st,gnum=gnum)
plotvolu2d(vd)

[Package denpro version 0.9.2 Index]