tree.segme {denpro}	R Documentation

Returns the segmentation of the nodes of a visualization tree

Description

Returns the segmentation of the nodes of a visualization tree. When the visualization tree is a tail tree, finds a grouping for the observations. The grouping may be used to enhance scatter plots, graphical matrices, and parallel coordinate plots.

Usage

tree.segme(tt, paletti = NULL, pcf = NULL)

Arguments

tt
visualization tree; for example a tail tree

paletti
a sequence of numbers or colors (group labels)

pcf
internal; piecewise constant function

Value

Vector of length n, where n is the number of nodes in the visualization tree. The elements of the vector give the segmentation labels.

Author(s)

Jussi Klemela

See Also

paracoor

Examples

dendat<-sim.data(n=1000,type="cross",seed=1)
rho<-1.1
tt<-leafsfirst(dendat=dendat,rho=rho)

ts<-tree.segme(tt)

plot(dendat,col=ts)

paracoor(dendat,paletti=ts)

[Package denpro version 0.9.2 Index]